% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbasummary_function.R
\name{cbaSummary}
\alias{cbaSummary}
\title{A Function to generate a summary table of costs and Benefits}
\usage{
cbaSummary(table)
}
\arguments{
\item{table}{The CBA table of the undiscounted and discounted costs and benefits over the appraisal period taken from the cbaTable() function}
}
\value{
An appraisal summary table giving the PVB, PVC, NPV and BCR values
}
\description{
This function reads in a table of discounted costs and benefits over the
selected appraisal period generated from the cbaTable function.
}
\keyword{cba}

