% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-benefit.R
\name{cost_benefit}
\alias{cost_benefit}
\title{Produce cost benefit summary table}
\usage{
cost_benefit(cost_table, time_benefits, veh_op_costs)
}
\arguments{
\item{cost_table}{The table of costs output by the cost_table() function.}

\item{time_benefits}{The table of time beneftis output by the 
time_benefits() function.}

\item{veh_op_costs}{The table of vehicle operating costs output by the 
veh_op_costs() function.}
}
\value{
A a summary of the cost benefit analysis results.
}
\description{
This function generates a summary of the cost benefit analysis results
including present value of costs (PVC), present value of benefits (PVB),
net present value (NPV) and the benefit to costs ratio (BCR.
}
\keyword{cba,}
\keyword{summary,}
\keyword{table}
