% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-table.R
\name{cost_table}
\alias{cost_table}
\title{Scheme cost profile}
\usage{
cost_table(cost_est, price_base_yr = 2011, opening_yr, appr_period = 30,
  disc_rate = 0.05, resid_period = 0, cpi_base = 99.4, cpi_cost_est,
  sppf = 1.3, spl = 1, labour_cont = 0.35, cost_yrs = c(2022:2025),
  cost_prop = c(0.116, 0.437, 0.409, 0.038))
}
\arguments{
\item{cost_est}{The cost estimate for the scheme,  entered in factor prices (millions)}

\item{price_base_yr}{The price base year assumed for the appraisal. Defaults to 2011}

\item{opening_yr}{The assumed opening year for the scheme}

\item{appr_period}{The period in years over which the scheme is being appraised. Defaults to 30.}

\item{disc_rate}{The test discount rate as a decimal.}

\item{resid_period}{The period in years, after the appraisal period which is assumed to represent the 
residual value of the scheme. Defaults to 0.}

\item{cpi_base}{The consumer price index for the price base year. Defaults tp 103.8 for 2011 (base Dec 2006)}

\item{cpi_cost_est}{The consumer price index for the month and year of the scheme cost estimate}

\item{sppf}{The shadow price of public funds. Defaults to 1.3}

\item{spl}{The shadow price of labour. Defaults to 1.0}

\item{cost_yrs}{The range of years over which the costs are assumed to be incurred}

\item{cost_prop}{a vector list of the proportion of scheme costs for each year definec in the costYears argument.}
}
\value{
A table of scheme costs per year to be input into the cbaTable() function
}
\description{
This function generates a table of costs broken down by type and year for input into a CBA
}
\keyword{cba,}
\keyword{costs}
