% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costTable_function.R
\name{costTable}
\alias{costTable}
\title{Scheme cost profile}
\usage{
costTable(costEst, priceBaseYear = 2011, openingYear, appraisalPeriod = 30,
  residualValuePeriod = 0, cpiBase = 103.8, cpiCostEst, sppf = 1.3,
  spl = 1, costYears = c(2022:2025), costProp = c(0.116, 0.437, 0.409,
  0.038))
}
\arguments{
\item{costEst}{The cost estimate for the scheme,  entered in factor prices (millions)}

\item{priceBaseYear}{The price base year assumed for the appraisal. Defaults to 2011}

\item{openingYear}{The assumed opening year for the scheme}

\item{appraisalPeriod}{The period in years over which the scheme is being appraised. Defaults to 30.}

\item{residualValuePeriod}{The period in years, after the appraisal period which is assumed to represent the 
residual value of the scheme. Defaults to 0.}

\item{cpiBase}{The consumer price index for the price base year. Defaults tp 103.8 for 2011 (base Dec 2006)}

\item{cpiCostEst}{The consumer price index for the month and year of the scheme cost estimate}

\item{sppf}{The shadow price of public funds. Defaults to 1.3}

\item{spl}{The shadow price of labour. Defaults to 1.0}

\item{costYears}{The range of years over which the costs are assumed to be incurred}

\item{costProp}{a vector list of the proportion of scheme costs for each year definec in the costYears argument.}
}
\value{
A table of scheme costs per year to be input into the cbaTable() function
}
\description{
This function generates a table of costs broken down by type and year for input into a CBA
}
\keyword{cba,}
\keyword{costs}

