% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel_cons_param}
\alias{fuel_cons_param}
\title{Fuel consumption parameters.}
\format{A data frame with five variables: \code{vehicle}: vehicle types,
 \code{a}, \code{b}, \code{c} and \code{d}: fuel consumption parameters.}
\usage{
fuel_cons_param
}
\description{
Fuel consumption parameters for use in TUBA based on WebTAG Table A1.3.8
}
\keyword{datasets}
