% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel_cost_2011}
\alias{fuel_cost_2011}
\title{2011 fuel costs.}
\format{A data frame with 2 variables: \code{fuel}, and \code{price}}
\usage{
fuel_cost_2011
}
\description{
Petrol and diesel fuel costs in cents from 2011.
}
\keyword{datasets}
