% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel-cost.R
\name{fuel_cost_km}
\alias{fuel_cost_km}
\title{Table of fuel consumption costs}
\usage{
fuel_cost_km(speed, fuel_cons_param, fuel_split, fuel_cost_2011)
}
\arguments{
\item{speed}{A vector of traffic speeds.}

\item{fuel_cons_param}{Fuel consumption parameters available in package}

\item{fuel_split}{2011 Fleet vehicle splits by fuel type available in 
package}

\item{fuel_cost_2011}{Petrol and diesel fuel costs in cents from 2011 available
in package}
}
\value{
A table of fuel consumption costs per km for the specified vector of
speeds
}
\description{
This function generates a table of fuel consumption costs per km for the 
specified vector of traffic speeds (in kph) for the standard vehicle and
fuel types.
}
\keyword{cba,}
\keyword{consumption}
\keyword{fuel,}
