% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel_split}
\alias{fuel_split}
\title{2011 vehicle splits.}
\format{A data frame with 3 variables: \code{vehicle}: vehicle types,
\code{petrol} and \code{diesel} splits}
\usage{
fuel_split
}
\description{
2011 Fleet vehicle splits by fuel type.
}
\keyword{datasets}
