% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pag_t532}
\alias{pag_t532}
\title{PAG link based growth factors}
\format{A data frame with 5 variables: \code{reg}: Geographical reiogn as
per Figure 5.3.1 of PAG Unit 5.3, \code{growth}: low/central/high growth
scenario, \code{veh}: light vehicle (lv) or heavy vehicle (hv),
\code{period}: period in years 2013-2030 (1330) or 2030-2050 (3050), and
\code{factor}: annual growth factor}
\usage{
pag_t532
}
\description{
PAG Table 5.3.2: Annual Growth Factors by region
}
\keyword{datasets}
