% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-benefits.R
\name{time_benefits}
\alias{time_benefits}
\title{Calculate time savings benefits over full appraisal period}
\usage{
time_benefits(opening_yr, appr_period = 30, resid_period = 30, disc_rate,
  price_base_yr = 2011, ave_veh_occ = 1.2, traffic_proj, time_saving)
}
\arguments{
\item{opening_yr}{The proposed year of scheme opening}

\item{appr_period}{The standard appraisal period in years}

\item{resid_period}{The residual value period in years (in
addition to the standard appraisal period)}

\item{disc_rate}{The test discount rate as a decimal}

\item{price_base_yr}{The price base year assumed for the 
appraisal. Defaults to 2011}

\item{ave_veh_occ}{The average vehicle occupancy (persions
per vehicle)}

\item{traffic_proj}{A dataframe of annual traffic flow projections}

\item{time_saving}{A vector of time savings in hours}
}
\value{
A dataframe of discounted scheme benefits
}
\description{
A function to read in appraisal paramaters, time savings 
calculated and traffic projections to calculate scheme
benefits from time savings
}
