% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-saving.R
\name{time_saving}
\alias{time_saving}
\title{Calculate time savings based on scheme lengths & speeds}
\usage{
time_saving(ex_length, prop_length, ex_speed, proj_speed)
}
\arguments{
\item{ex_length}{The length (in km) of the existing road section}

\item{prop_length}{The length (in km) of the prposed scheme}

\item{ex_speed}{Current average speeds on the existing route}

\item{proj_speed}{Projected average speeds on upgraded route}
}
\value{
A vector of time savings in hours
}
\description{
This function reads in outline lengths and average speeds
for the existing section of road and proposed scheme and
estimates time savings per vehicle
}
