% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSavings_function.R
\name{timeSavings}
\alias{timeSavings}
\title{Simple time savings estimate}
\usage{
timeSavings(dmLength, dsLength, dmSpeedLimit, dsSpeedLimit, openAadt,
  forecastAadt)
}
\arguments{
\item{dmLength}{The length of road section in the 'Do Minumum' scenario (i.e without scheme)}

\item{dsLength}{The length of road section in the 'Do Something scenaio (i.e with scheme)}

\item{dmSpeedLimit}{The existing speed limit (kph) on the road section}

\item{dsSpeedLimit}{The proposed speed limit (kph)}

\item{openAadt}{The estimated scheme AADT in the opening year}

\item{forecastAadt}{The estimated scheme AADT in the forecast year}
}
\description{
A function to estimate time savings from a simple road upgrade scheme
}
\keyword{cba}

