% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic-proj.R
\name{traffic_proj}
\alias{traffic_proj}
\title{Output list of traffic flow projections by year}
\usage{
traffic_proj(base_yr, base_aadt, pc_hgv, opening_yr, region)
}
\arguments{
\item{base_yr}{The year of observed traffic volumes}

\item{base_aadt}{Base year observed traffic volumes as AADT}

\item{opening_yr}{The proposed year of scheme opening}

\item{region}{The region in which the scheme is located as per 
PAG Unit 5.3}
}
\value{
A dataframe of annual traffic flow projections
}
\description{
This function reads in base year AADT and the region in which
the scheme is located and outputs a list of traffic flow
projections based on PAG Unit 5.3 link based projections
}
