% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veh-op-costs.R
\name{veh_op_costs}
\alias{veh_op_costs}
\title{Calculate vehicle operating costs over full appraisal period}
\usage{
veh_op_costs(opening_yr, appr_period = 30, resid_period = 30, disc_rate,
  price_base_yr = 2011, ave_veh_occ = 1.2, traffic_proj, speed_ex,
  speed_prop, fuel_costs, non_fuel_costs)
}
\arguments{
\item{opening_yr}{The proposed year of scheme opening}

\item{appr_period}{The standard appraisal period in years}

\item{resid_period}{The residual value period in years (in
addition to the standard appraisal period)}

\item{disc_rate}{The test discount rate as a decimal}

\item{price_base_yr}{The price base year assumed for the 
appraisal. Defaults to 2011}

\item{ave_veh_occ}{The average vehicle occupancy (persions
per vehicle)}

\item{traffic_proj}{A dataframe of annual traffic flow projections}

\item{speed_ex}{Average speed on existing route in kph}

\item{speed_prop}{Average speed on proposed route in kph}

\item{fuel_costs}{Table of fuel costs generated by fuel_costs_km function}

\item{non_fuel_costs}{Table of non-fuel costs generated by non_fuel_cost_km
function}
}
\value{
A dataframe of discounted scheme benefits
}
\description{
A function to read in appraisal paramaters, fuel and non-fuel costs 
calculated and traffic projections to calculate scheme vehicle operating
costs
}
