% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaTable_function.R
\name{cbaTable}
\alias{cbaTable}
\title{Generates a table of costs and Benefits}
\usage{
cbaTable(costTable, appraisalPeriod, residualValuePeriod, openingYear,
  forecastYear, timeSavings, aveVoT, discountRate, priceBaseYear)
}
\arguments{
\item{appraisalPeriod}{The number of years over which the scheme should be appraised}

\item{residualValuePeriod}{The numnber of years beyond the appraisal period where the scheme benefits can be taken as a residual value}

\item{openingYear}{The proposed opening year for the scheme being appraised}

\item{forecastYear}{A future year for which where scheme impacts have been forecast}

\item{timeSavings}{The estimated time savings (in hours) due to the scheme in the opening year and forecast year
as a list, (i.e. c(x, y)). Can also use output by the timeSavings() function.}

\item{aveVoT}{An average value of time assumed for the calculation of scheme benefits}

\item{discountRate}{The project discount rate used to convert future year prices to the price base year}

\item{priceBaseYear}{The year that all appraisal prices are to be represented in}

\item{capCosts}{The capital costs of the proposed scheme}

\item{costsYear}{The year in which capital costs are incurred (select single year for simplicity)}
}
\value{
A table of the undiscounted and discounted costs and benefits over the appraisal period.
}
\description{
This takes standard appraisal parameters plus time savings  and costs estimated elsewhere
to produce a table of discounted costs and benefits over the selected appraisal period.
}
\keyword{cba}

